// src/types/perspectiveProvider.ts
export type PerspectiveProviderFields = {
  pincode: string
  zone: string
  providerName: string
  contactPersonName: string
  email: string
  contactNo: string
  requestNumber: string
  district: string
  state: string
  parentDbId: string
}

export enum EEmpanelmentType {
  Auto = 'Auto',
  Manual = 'Manual',
}